requirements.txt
fastapi[standard]
PyJWT
cryptography
python-dotenv